+++ nacc_attribute_deriver/attributes/mqt/study_parameters.py
@@ -18,19 +18,21 @@ class StudyParametersAttributeCollection(AttributeCollection):
     """Class to collect study-parameter attributes."""
 
     def __init__(self, table: SymbolTable):
-        self.__file = UDSNamespace(table)
-        self.__subject_info = SubjectInfoNamespace(table)
+        self.__uds = UDSNamespace(table)
+        self.__subject_info = SubjectInfoNamespace(table=table)
 
     def _create_uds_versions_available(self) -> List[str]:
         """Keeps track of available UDS versions."""
-        formver = self.__file.normalized_formver()
-        versions = self.__subject_info.get_value("study-parameters.uds.versions", [])
-        versions = {
-            version
-            for version in versions
-            if isinstance(version, str) and re.match(r"UDSv[1-4]", version)
-        }
-        if formver:
-            versions.add(f"UDSv{formver}")
+        formver = self.__uds.normalized_formver()
+        versions = set([f"UDSv{formver}"])
 
-        return list(versions)
+        prev_versions = self.__subject_info.get_value(
+            "study-parameters.uds.versions", list, default=[]
+        )
+
+        if prev_versions:
+            for v in prev_versions:
+                if isinstance(v, str) and re.match(r"UDSv[1-4]", v):
+                    versions.add(v)
+
+        return sorted(list(versions))