+++ nacc_attribute_deriver/attributes/nacc/modules/mds/form_mds.py
@@ -10,24 +10,27 @@ from nacc_attribute_deriver.utils.date import create_death_date
 
 class MDSFormAttributeCollection(AttributeCollection):
     def __init__(self, table: SymbolTable) -> None:
-        self.__mds = FormNamespace(table)
+        self.__mds = FormNamespace(table=table, required=frozenset(["vitalst"]))
+
+        self.__vitalst = self.__mds.get_required("vitalst", int)
 
     def _create_mds_death_date(self) -> Optional[date]:
-        if not self.is_target_int(self.__mds.get_value("vitalst"), 2):
+        """MDS death date; can be unknown."""
+        if self.__vitalst != 2:
             return None
 
-        year = self.__mds.get_value("deathyr")  # can be 9999
-        month = self.__mds.get_value("deathmo")  # can be 99
-        day = self.__mds.get_value("deathday")  # can be 99
+        year = self.__mds.get_value("deathyr", int)  # can be 9999
+        month = self.__mds.get_value("deathmo", int)  # can be 99
+        day = self.__mds.get_value("deathday", int)  # can be 99
 
         return create_death_date(year=year, month=month, day=day)
 
     def _create_mds_death_month(self) -> Optional[int]:
         """MDS death month - can be 99."""
-        if not self.is_target_int(self.__mds.get_value("vitalst"), 2):
+        if self.__vitalst != 2:
             return None
 
-        month = self.__mds.get_value("deathmo")
+        month = self.__mds.get_value("deathmo", int)
 
         try:
             if month is not None:
@@ -38,4 +41,5 @@ class MDSFormAttributeCollection(AttributeCollection):
         return month if month is not None else 99
 
     def _create_mds_vitalst(self) -> int:
-        return self.__mds.get_value("vitalst")
+        """MDS VITALST."""
+        return self.__vitalst