+++ nacc_attribute_deriver/attributes/nacc/modules/np/form_np.py
@@ -18,13 +18,16 @@ from .np_mapper import NPMapper
 class NPFormAttributeCollection(AttributeCollection):
     def __init__(self, table: SymbolTable) -> None:
         """Check that this is an NP form."""
-        self.__np = FormNamespace(table)
+        self.__np = FormNamespace(
+            table=table, required=frozenset(["formver", "module"])
+        )
 
-        module = self.__np.get_value("module")
-        if not module or module.upper() != "NP":
+        module = self.__np.get_required("module", str)
+        if module.upper() != "NP":
             msg = f"Current file is not an NP form: found {module}"
             raise InvalidFieldError(msg)
 
+        self.formver: int = self.__np.get_required("formver", int)
         self.mapper = NPMapper(self.__np)
         self.form_evaluator = NPFormWideEvaluator(self.__np, self.mapper)
 
@@ -33,15 +36,14 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Cerebral amyloid angiopathy
         """
-        formver = self.__np.get_value("formver")
-        npamy = self.__np.get_value("npamy")
+        npamy = self.__np.get_value("npamy", int)
         naccamy = npamy
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             pass
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             naccamy = self.mapper.map_sub1(npamy)
-        elif formver == 1:
+        elif self.formver == 1:
             naccamy = (
                 self.mapper.map_sub1(npamy) if npamy else self.mapper.map_vasc(naccamy)
             )
@@ -54,19 +56,23 @@ class NPFormAttributeCollection(AttributeCollection):
         Severity of gross findings — atherosclerosis of the circle of
         Willis
         """
-        formver = self.__np.get_value("formver")
-        npavas = self.__np.get_value("npavas")
+        npavas = self.__np.get_value("npavas", int)
         naccavas = npavas
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             pass
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
+            if not npavas:
+                raise InvalidFieldError(
+                    "npvas cannot be missing for formver 7, 8, or 9"
+                )
+
             naccavas = npavas - 1
             if npavas == 5:
                 naccavas = 8
             elif npavas == 9:
                 naccavas = 9
-        elif formver == 1:
+        elif self.formver == 1:
             naccavas = (
                 self.mapper.map_sub1(npavas)
                 if npavas
@@ -90,16 +96,15 @@ class NPFormAttributeCollection(AttributeCollection):
         TODO: QAF does have this as -4 in some cases, but RDD does not say
         it can be -4 - probably best to keep as 9 but flagging
         """
-        formver = self.__np.get_value("formver")
-        npcort = self.__np.get_value("npcort")
-        npftdtau = self.__np.get_value("npftdtau")
+        npcort = self.__np.get_value("npcort", int)
+        npftdtau = self.__np.get_value("npftdtau", int)
         nacc_cbd = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             nacc_cbd = self.mapper.map_v10(npcort, npftdtau)
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             nacc_cbd = self.mapper.map_v9(npcort)
-        elif formver == 1:
+        elif self.formver == 1:
             nacc_cbd = (
                 self.mapper.map_v9(npcort) if npcort else self.mapper.map_vasc(nacc_cbd)
             )
@@ -111,15 +116,14 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Density of diffuse plaques (CERAD semi-quantitative score)
         """
-        formver = self.__np.get_value("formver")
-        npdiff = self.__np.get_value("npdiff")
+        npdiff = self.__np.get_value("npdiff", int)
         naccdiff = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             naccdiff = npdiff
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             naccdiff = self.mapper.map_sub4(npdiff)
-        elif formver == 1:
+        elif self.formver == 1:
             naccdiff = (
                 self.mapper.map_sub4(npdiff)
                 if npdiff
@@ -133,10 +137,9 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Down syndrome
         """
-        formver = self.__np.get_value("formver")
-        npchrom = self.__np.get_value("npchrom")
+        npchrom = self.__np.get_value("npchrom", int)
 
-        if formver in [10, 11] or formver in [7, 8, 9] or formver == 1:
+        if self.formver in [10, 11] or self.formver in [7, 8, 9] or self.formver == 1:
             np_down = 1 if npchrom == 11 else 7
 
         return np_down
@@ -148,18 +151,17 @@ class NPFormAttributeCollection(AttributeCollection):
 
         infarcts or lacunes
         """
-        formver = self.__np.get_value("formver")
-        npinf = self.__np.get_value("npinf")
-        nplinf = self.__np.get_value("nplinf")
-        nplac = self.__np.get_value("nplac")
+        npinf = self.__np.get_value("npinf", int)
+        nplinf = self.__np.get_value("nplinf", int)
+        nplac = self.__np.get_value("nplac", int)
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             return npinf if npinf is not None else 9
 
-        if formver in [7, 8, 9]:
+        if self.formver in [7, 8, 9]:
             return self.mapper.map_comb2(nplinf, nplac)
 
-        if formver == 1 and (nplinf is not None and nplac is not None):
+        if self.formver == 1 and (nplinf is not None and nplac is not None):
             return self.mapper.map_comb2(nplinf, nplac)
 
         return 9  # Fallback value
@@ -171,18 +173,17 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Laminar necrosis
         """
-        formver = self.__np.get_value("formver")
-        npnec = self.__np.get_value("npnec")
-        nppath = self.__np.get_value("nppath")
-        npgross = self.__np.get_value("npgross")
-        npvasc = self.__np.get_value("npvasc")
+        npnec = self.__np.get_value("npnec", int)
+        nppath = self.__np.get_value("nppath", int)
+        npgross = self.__np.get_value("npgross", int)
+        npvasc = self.__np.get_value("npvasc", int)
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             return self.mapper.map_v10(npnec, nppath)
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             return self.mapper.map_v9(npnec)
 
-        elif formver == 1:
+        elif self.formver == 1:
             if npgross == 2 or npvasc == 2:
                 return 0
             elif npgross == 9:
@@ -201,13 +202,13 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Other pathologic diagnosis.
         """
-        formver = self.__np.get_value("formver")
-        nppdxr = self.__np.get_value("nppdxr")
-        nppdxs = self.__np.get_value("nppdxs")
-        nppdxt = self.__np.get_value("nppdxt")
+        nppdxr = self.__np.get_value("nppdxr", int)
+        nppdxs = self.__np.get_value("nppdxs", int)
+        nppdxt = self.__np.get_value("nppdxt", int)
+        npmajor = self.__np.get_value("npmajor", int)
         naccothp = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             if nppdxr == 1 or nppdxs == 1 or nppdxt == 1:
                 return 1
             elif nppdxr == 0 and nppdxs == 0 and nppdxt == 0:
@@ -215,14 +216,14 @@ class NPFormAttributeCollection(AttributeCollection):
             else:
                 return 9
 
-        elif formver in [7, 8, 9]:
-            return self.mapper.map_v9(self.__np.get_value("npmajor"))
+        elif self.formver in [7, 8, 9]:
+            return self.mapper.map_v9(npmajor)
 
-        elif formver == 1:
-            if self.__np.get_value("npmajor") is not None:
-                naccothp = self.mapper.map_v9(self.__np.get_value("npmajor"))
+        elif self.formver == 1:
+            if npmajor is not None:
+                naccothp = self.mapper.map_v9(npmajor)
             else:
-                naccothp = self.mapper.map_gross(self.__np.get_value("naccothp"))
+                naccothp = self.mapper.map_gross(naccothp)
 
         return naccothp if naccothp is not None else 9
 
@@ -234,24 +235,17 @@ class NPFormAttributeCollection(AttributeCollection):
         TODO: QAF does have this as -4 in some cases, but RDD does not say
         it can be -4 - probably best to keep as 9 but flagging
         """
-        formver = self.__np.get_value("formver")
-        npftdtau = self.__np.get_value("npftdtau")
-        nppick = self.__np.get_value("nppick")
+        npftdtau = self.__np.get_value("npftdtau", int)
+        nppick = self.__np.get_value("nppick", int)
         naccpick = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             return self.mapper.map_v10(nppick, npftdtau)
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             return self.mapper.map_v9(nppick)
 
-        elif formver == 1:
-            npgross = self.__np.get_value("npgross")
-            if npgross == 2:
-                return 0
-            elif npgross == 9:
-                return 9
-            else:
-                naccpick = self.mapper.map_gross(nppick)
+        elif self.formver == 1:
+            naccpick = self.mapper.map_gross(nppick)
 
         return naccpick if naccpick is not None else 9
 
@@ -260,18 +254,17 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Prion disease.
         """
-        formver = self.__np.get_value("formver")
-        npcj = self.__np.get_value("npcj")
-        npprion = self.__np.get_value("npprion")
-        nppdxc = self.__np.get_value("nppdxc")
+        npcj = self.__np.get_value("npcj", int)
+        npprion = self.__np.get_value("npprion", int)
+        nppdxc = self.__np.get_value("nppdxc", int)
         naccprio = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             return nppdxc if nppdxc is not None else 9
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             return self.mapper.map_comb2(npcj, npprion)
 
-        elif formver == 1:
+        elif self.formver == 1:
             if npcj is not None and npprion is not None:
                 return self.mapper.map_comb2(npcj, npprion)
 
@@ -284,24 +277,17 @@ class NPFormAttributeCollection(AttributeCollection):
 
         FTLD-tau subtype — progressive supranuclear palsy (PSP).
         """
-        formver = self.__np.get_value("formver")
-        npftdtau = self.__np.get_value("npftdtau")
-        npprog = self.__np.get_value("npprog")
-        npgross = self.__np.get_value("npgross")
+        npftdtau = self.__np.get_value("npftdtau", int)
+        npprog = self.__np.get_value("npprog", int)
         naccprog = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             return self.mapper.map_v10(npprog, npftdtau)
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             return self.mapper.map_v9(npprog)
 
-        elif formver == 1:
-            if npgross == 2:
-                return 0
-            elif npgross == 9:
-                return 9
-            else:
-                naccprog = self.mapper.map_gross(npprog)
+        elif self.formver == 1:
+            self.mapper.map_gross(npprog)
 
         return naccprog if naccprog is not None else 9
 
@@ -320,19 +306,14 @@ class NPFormAttributeCollection(AttributeCollection):
             num: Number of NPMPATHX variable to grab which corresponds
                 to the number of the NACCWRIX variable
         """
-        formver = self.__np.get_value("formver")
-        writein = self.__np.get_value(writein)
-        npmpathx = self.__np.get_value(f"npmpath{num}")
+        value = self.__np.get_value(writein, str)
+        npmpathx = self.__np.get_value(f"npmpath{num}", str)
         naccwrix = None
 
-        if formver in [10, 11]:
-            naccwrix = writein
-        elif formver in [1, 7, 8, 9]:
+        if self.formver in [10, 11]:
+            naccwrix = value
+        elif self.formver in [1, 7, 8, 9]:
             naccwrix = npmpathx
-
-        # strip whitespace since write-in
-        if naccwrix is not None:
-            naccwrix = naccwrix.strip()
 
         return naccwrix if naccwrix else None
 
@@ -362,16 +343,15 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Braak stage for neurofibrillary degeneration (B score)
         """
-        formver = self.__np.get_value("formver")
-        npbraak = self.__np.get_value("npbraak")
+        npbraak = self.__np.get_value("npbraak", int)
         naccbraa = npbraak
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             pass
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             if npbraak == 7:
                 naccbraa = 0
-        elif formver == 1:
+        elif self.formver == 1:
             if npbraak:
                 if npbraak == 7:
                     naccbraa = 0
@@ -385,15 +365,14 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Density of neocortical neuritic plaques (CERAD score) (C score)
         """
-        formver = self.__np.get_value("formver")
-        npneur = self.__np.get_value("npneur")
+        npneur = self.__np.get_value("npneur", int)
         naccneur = npneur
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             pass
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             naccneur = self.mapper.map_sub4(npneur)
-        elif formver == 1:
+        elif self.formver == 1:
             naccneur = (
                 self.mapper.map_sub4(npneur)
                 if npneur
@@ -407,16 +386,15 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Microinfarcts
         """
-        formver = self.__np.get_value("formver")
-        npold = self.__np.get_value("npold")
-        npmicro = self.__np.get_value("npmicro")
+        npold = self.__np.get_value("npold", int)
+        npmicro = self.__np.get_value("npmicro", int)
 
         naccmicr = npold
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             pass
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             naccmicr = self.mapper.map_v9(npmicro)
-        elif formver == 1:
+        elif self.formver == 1:
             naccmicr = (
                 self.mapper.map_v9(npmicro)
                 if npmicro
@@ -430,13 +408,12 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Hemorrhages and microbleeds
         """
-        formver = self.__np.get_value("formver")
-        nphem = self.__np.get_value("nphem")
+        nphem = self.__np.get_value("nphem", int)
         nacchem = None
 
-        if formver in [10, 11]:
-            nphemo = self.__np.get_value("nphemo")
-            npoldd = self.__np.get_value("npoldd")
+        if self.formver in [10, 11]:
+            nphemo = self.__np.get_value("nphemo", int)
+            npoldd = self.__np.get_value("npoldd", int)
 
             if nphemo == 1 or npoldd == 1:
                 nacchem = 1
@@ -446,10 +423,10 @@ class NPFormAttributeCollection(AttributeCollection):
                 nacchem = 8
             else:
                 nacchem = 9
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             nacchem = nphem
             nacchem = self.mapper.map_v9(nphem)
-        elif formver == 1:
+        elif self.formver == 1:
             nacchem = nphem
             nacchem = (
                 self.mapper.map_v9(nphem) if nphem else self.mapper.map_vasc(nacchem)
@@ -462,15 +439,14 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Arteriolosclerosis
         """
-        formver = self.__np.get_value("formver")
-        nparter = self.__np.get_value("nparter")
+        nparter = self.__np.get_value("nparter", int)
         naccarte = None
 
-        if formver in [10, 11]:
+        if self.formver in [10, 11]:
             naccarte = nparter
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             naccarte = self.mapper.map_sub1(nparter)
-        elif formver == 1:
+        elif self.formver == 1:
             naccarte = (
                 self.mapper.map_sub1(nparter)
                 if nparter
@@ -484,20 +460,19 @@ class NPFormAttributeCollection(AttributeCollection):
 
         Lewy body disease
         """
-        formver = self.__np.get_value("formver")
         nacclewy = None
 
-        if formver in [10, 11]:
-            nplbod = self.__np.get_value("nplbod")
+        if self.formver in [10, 11]:
+            nplbod = self.__np.get_value("nplbod", int)
             nacclewy = nplbod
             if nplbod == 4:
                 nacclewy = 2
             if nplbod == 5:
                 nacclewy = 4
-        elif formver in [7, 8, 9]:
+        elif self.formver in [7, 8, 9]:
             nacclewy = self.mapper.map_lewy()
-        elif formver == 1:
-            nplewy = self.__np.get_value("nplewy")
+        elif self.formver == 1:
+            nplewy = self.__np.get_value("nplewy", int)
             if nplewy:
                 nacclewy = self.mapper.map_lewy()
             else:
@@ -507,14 +482,14 @@ class NPFormAttributeCollection(AttributeCollection):
         return nacclewy if nacclewy is not None else 9
 
     def _create_np_death_age(self) -> Optional[int]:
-        return self.__np.get_value("npdage")
+        return self.__np.get_value("npdage", int)
 
     def _create_np_death_date(self) -> Optional[date]:
-        if self.__np.get_value("npdage") is None:
+        if self.__np.get_value("npdage", int) is None:
             return None
 
-        year = self.__np.get_value("npdodyr")
-        month = self.__np.get_value("npdodmo")
-        day = self.__np.get_value("npdoddy")
+        year = self.__np.get_value("npdodyr", int)
+        month = self.__np.get_value("npdodmo", int)
+        day = self.__np.get_value("npdoddy", int)
 
         return create_death_date(year=year, month=month, day=day)