+++ tests/attributes/base/test_attributes.py
@@ -1,7 +1,4 @@
-import pytest
-
 from nacc_attribute_deriver.attributes.base.namespace import BaseNamespace
-from nacc_attribute_deriver.schema.errors import MissingRequiredError
 from nacc_attribute_deriver.symbol_table import SymbolTable
 
 
@@ -12,18 +9,7 @@ class TestAttribute:
         table["test.value"] = 5
         table["some.other.prefix.value"] = 10
 
-        attr = BaseNamespace(table, attribute_prefix="test.")
-        assert attr.get_value("value") == 5
-        assert attr.get_value("missing") is None
-        assert attr.get_value("missing", "default") == "default"
-
-    def test_assert_required(self):
-        """Test the assert_required method."""
-        table = SymbolTable()
-        table["file.info.forms.json"] = {}
-        attr = BaseNamespace(table, "file.info.forms.json.")
-
-        # should raise error since there's nothing int able
-        with pytest.raises(MissingRequiredError) as error:
-            attr.assert_required(["testvar"])
-        assert error.value.field == "file.info.forms.json.testvar"
+        attr = BaseNamespace(table=table, attribute_prefix="test.")
+        assert attr.get_value("value", int) == 5
+        assert attr.get_value("missing", int) is None
+        assert attr.get_value("missing", str, default="default") == "default"