+++ tests/attributes/mqt/test_study_parameters.py
@@ -11,7 +11,20 @@ from nacc_attribute_deriver.symbol_table import SymbolTable
 def table() -> SymbolTable:
     """Create dummy data and return it in an attribute object."""
     data = {
-        "file": {"info": {"forms": {"json": {"formver": "4", "module": "UDS"}}}},
+        "file": {
+            "info": {
+                "forms": {
+                    "json": {
+                        "formver": "4",
+                        "module": "UDS",
+                        "packet": "F",
+                        "birthmo": 1,
+                        "birthyr": 1950,
+                        "visitdate": "2025-01-10",
+                    }
+                }
+            }
+        },
         "subject": {"info": {"study-parameters": {"uds": {"versions": ["UDSv2"]}}}},
     }
 
@@ -21,5 +34,5 @@ def table() -> SymbolTable:
 class TestStudyParametersAttributeCollection:
     def test_create_uds_versions_available(self, table):
         """Tests _create_uds_versions_available."""
-        attr = StudyParametersAttributeCollection.create(table)
+        attr = StudyParametersAttributeCollection(table)
         assert set(attr._create_uds_versions_available()) == {"UDSv4", "UDSv2"}